Turn block into a closure in test macro
The `doc_test!` macro defines two local variables (`entries` and `doc`),
which are then accessed from the provided `block`.

However, this only compiled due to a bug in rustc: substituted
metavariables cannot refer to names defined within the macro body. For
example, the following ode does not compile:

```rust
macro_rules! foo {
    ($block:expr) => {
        let mut bar = false;
        $block
    }
}

fn main() {
    foo!({bar = true});
}
```

In this case, the `doc_test!` macro was incorrectly allowed to compile
due to the presence of the `#[tokio::test]` macro on the enclosing
function. When the underlying compiler bug is fixed in
https://github.com/rust-lang/rust/pull/75800,
this macro will stop compiling.

I've adjusted the macro to take in a closure instead of a block. This
closure is invoked with the `entries` and `doc` variables, allowing it
to compile with both old and new compilers.

If you have any questions about this change, feel free to ask me.